pyproject.toml
[project]
name = "sophie-sales-management-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic==1.16.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "bcrypt==4.3.0",
    "cffi==1.17.1",
    "click==8.1.8",
    "cryptography==45.0.3",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.115.12",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "idna==3.10",
    "jinja2>=3.1.6",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "passlib==1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "python-dotenv==1.1.0",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "requests>=2.32.3",
    "rsa==4.9.1",
    "six==1.17.0",
    "slowapi==0.1.9",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "starlette==0.46.2",
    "strawberry-graphql[fastapi]>=0.270.5",
    "tomli==2.2.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "uvicorn==0.34.3",
]
